refator mtdnn tasks
